Ensure the transcript is not referenced by another transcript message before deleting the attachments

+++ MixinServices/MixinServices/Foundation/File Management/AttachmentContainer.swift
@@ -66,7 +66,12 @@ public enum AttachmentContainer {
         }
     }
     
-    public static func removeAll(conversationId: String, transcriptId: String) {
+    public static func removeAll(transcriptId: String) {
+        guard !TranscriptMessageDAO.shared.hasTranscriptMessage(withMessageId: transcriptId) else {
+            // This situation reveals that this transcript message is referenced by another
+            // transcript message, in other words, it has a nested structure.
+            return
+        }
         let url = Self.url(transcriptId: transcriptId, filename: nil)
         try? FileManager.default.removeItem(at: url)
     }
